requirements.txt
# ── API Framework ──────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── LLM & Agent Orchestration ─────────────────────────────
langchain==0.3.7
langchain-google-genai==2.0.4
langgraph==0.2.53

# ── Database ──────────────────────────────────────────────
snowflake-connector-python==3.12.3

# ── PDF Processing ────────────────────────────────────────
PyMuPDF==1.24.11

# ── Data Validation & Config ─────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
